requirement.txt
argparse==1.2.1
Django==1.5
django-haystack==2.1.0
django-mptt==0.5
gunicorn==0.17.2
MySQL-python==1.2.4
wsgiref==0.1.2
xapian-haystack==2.0.0
django-notifications-hq
django-upload-avatar
